nyota-enterprise/orchestrator/requirements.txt
temporalio==1.5.0
nats-py==2.2.0
pydantic==2.5.2
httpx==0.25.2
